Data Cleaning Project
In this project we were provided with a dataset on Nashville housing that contained duplicate data, wrong date formats, and null values. I have went through the data using SQL to clean the data and correct some of the issues that were found. Feel free to look through the query and test it to see how the data changes throughout the project.